Context works
As of this commit, the post list is filtered a) when the page loads via useEffect (based on the context of catagory), and b) when when the filter is interacted with by calling the function. Is it better for the useEffect to watch the context and never call the function otherwise?